moved eventcard test files
