+++ src/components/EventCards/__tests__/CreateEventCard.test.js
@@ -4,9 +4,9 @@ import { MockedProvider } from '@apollo/client/testing';
 import { LocalizationProvider } from '@mui/x-date-pickers/LocalizationProvider';
 import { AdapterMoment } from '@mui/x-date-pickers/AdapterMoment';
 
-import CreateEventCard from '../components/EventCards/CreateEventCard';
-import { LoggableEventsContext } from '../providers/LoggableEventsProvider';
-import { ViewOptionsContext } from '../providers/ViewOptionsProvider';
+import CreateEventCard from '../CreateEventCard';
+import { LoggableEventsContext } from '../../../providers/LoggableEventsProvider';
+import { ViewOptionsContext } from '../../../providers/ViewOptionsProvider';
 
 describe('CreateEventCard', () => {
     const mockEventLabels = [
